(PUP-3008) Guard constant initialization conflict.
win32-service and win32-eventlog both define module Windows::Constant with the members
	ERROR_INSUFFICIENT_BUFFER, INFINITE, WAIT_FAILED, FORMAT_MESSAGE_IGNORE_INSERTS
When both modules are required, as in running puppet agent, a warning is given about
re-initializing already initialized constants. This patch applies the simplest fix
of using '||=' to guard constant initialization, on the condition that none of the
constants initialize to false. A better fix is being discussed with the gem owner.